Remove AUTOINCREMENT from primary key
As described in https://www.sqlite.org/autoinc.html, the AUTOINCREMENT
keyword imposes overhead and should be avoided if not needed. The
keyword will "prevent the reuse of ROWIDs over the lifetime of the
database", which is not a requirement for this app. Without the
keyword, the row id will still be automatically filled - it's just that
it may use the id of a row that was previously deleted, which is fine.