Switch from SQLite to firestore
With this commit, running the app locally works (using the Firestore
emulator described in the README) and unit tests pass. It also works
in prod since Google App Engine takes care of authorization (the
project service account has access).

Some changes will need to be made in the GitHub Actions build to run
the emulator.