Fix Firestore authentication when running on Actions
The app/tests were working locally but failing on Actions, because
the way I had set up authentication for Firestore, it was falling back
to the Google Cloud application default credentials, which were set
locally but not on Actions. Those creds are not needed, though, because
I'm using the Firestore emulator which runs alongside the app/tests.

The fix was to explicitly set the projects/creds to dev/anonymous when
not running on prod. On prod, everything is automatically inferred (but
that inference doesn't happen if you have set the projects/creds
yourself).